test: Morph into unit tests
Morph the judge configuration tests into unit tests.

Instead of verifying the Spring test execution listener behavior directly, the unit test verifies the used components.
